+++ lib/isolate_current_directory.dart
@@ -29,6 +29,17 @@ FileStat _statSync(String dir, String path) {
   return FileStat.statSync(absPath(path, dir));
 }
 
+T _invoke<T>(Function onError, Object e, StackTrace stackTrace) {
+  if (onError is Function(dynamic)) {
+    return onError(e);
+  }
+  if (onError is Function(dynamic, dynamic)) {
+    return onError(e, stackTrace);
+  }
+  // let Dart fail if this function can't be called like this
+  return onError();
+}
+
 /// Run the given [action] using the given [directory] as the working directory.
 ///
 /// Within the scope of [action], any [File], [Directory] and [Link]
@@ -46,12 +57,40 @@ FileStat _statSync(String dir, String path) {
 /// so modifications are only visible within [action]'s scope. Notice that
 /// the global value can still be accessed via [Zone.root], so this feature
 /// cannot be used for security.
+///
+/// If the [onError] callback is provided, it has similar semantics to the
+/// [Future.then] method's argument with the same name.
+/// In other words: it must accept one or two arguments,
+/// the first one being the error, and the second one, the stackTrace. It must
+/// return a value compatible with the return value of [action]. If it throws,
+/// this function throws the error, or if [action] is async, the returned
+/// [Future] completes with that error.
 FutureOr<T> withCurrentDirectory<T>(
-    String directory, FutureOr<T> Function() action) {
+    String directory, FutureOr<T> Function() action,
+    {Function? onError}) {
   final parentZone = Zone.current;
   var zoneDir = Directory(absPath(directory));
 
-  return IOOverrides.runZoned(() => action(),
+  return IOOverrides.runZoned(
+      () {
+        try {
+          final result = action();
+          if (result is Future<T>) {
+            if (onError == null) return result;
+            return result.catchError(onError);
+          }
+          return result;
+        } catch (e, stackTrace) {
+          if (onError == null) {
+            rethrow;
+          }
+          final result = _invoke(onError, e, stackTrace);
+          if (identical(e, result)) {
+            rethrow;
+          }
+          return result;
+        }
+      },
       createDirectory: (p) => IsolatedDirectory.of(p, parentZone),
       createFile: (p) => IsolatedFile.of(p, parentZone),
       createLink: (p) => IsolatedLink.of(p, parentZone),
@@ -77,7 +116,8 @@ FutureOr<T> withCurrentDirectory<T>(
 /// // AFTER
 /// Isolate.run(wrapWithCurrentDirectory(readTextFile));
 /// ```
-FutureOr<T> Function() wrapWithCurrentDirectory<T>(FutureOr<T> Function() fun) {
+FutureOr<T> Function() wrapWithCurrentDirectory<T>(FutureOr<T> Function() fun,
+    {Function? onError}) {
   final workingDir = Directory.current.path;
-  return () => withCurrentDirectory(workingDir, fun);
+  return () => withCurrentDirectory(workingDir, fun, onError: onError);
 }